# apps/api/docs/swagger.yaml
definitions:
  database.Challenge:
    properties:
      active:
        type: boolean
      config_json:
        type: string
      created_at:
        type: string
      description:
        type: string
      difficulty:
        type: integer
      duration_ms:
        type: integer
      id:
        type: integer
      name:
        type: string
      seed_type:
        type: string
      updated_at:
        type: string
    required:
    - config_json
    - difficulty
    - duration_ms
    - name
    - seed_type
    type: object
  database.Rank:
    properties:
      division:
        $ref: '#/definitions/sql.NullInt64'
      id:
        type: integer
      name:
        type: string
      tier_number:
        $ref: '#/definitions/sql.NullInt64'
    required:
    - name
    type: object
  database.Score:
    properties:
      challenge_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      score:
        type: integer
    required:
    - score
    type: object
  sql.NullInt64:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/challenge:
    get:
      description: Returns a single challenge by its unique numeric ID. Includes the
        full challenge configuration JSON (target layout, timing, etc.), difficulty,
        seed type, and active status. Returns 404 if the ID does not exist.
      parameters:
      - description: Unique challenge ID
        example: 1
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Challenge found
          schema:
            $ref: '#/definitions/database.Challenge'
        "400":
          description: Invalid or non-numeric challenge ID provided
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No challenge found with the given ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a challenge by ID
      tags:
      - challenges
  /api/v1/challenges:
    get:
      description: Returns a list of all available Airdribble training challenges.
        Each challenge includes its configuration (targets, duration, seed type),
        difficulty rating, and active status. Use this to populate the challenge selection
        screen.
      produces:
      - application/json
      responses:
        "200":
          description: List of challenges
          schema:
            items:
              $ref: '#/definitions/database.Challenge'
            type: array
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all challenges
      tags:
      - challenges
  /api/v1/challenges/{challenge_id}/session:
    post:
      description: Creates a new challenge session for the currently authenticated
        guest user. This must be called before the user starts a challenge attempt.
        On success, a short-lived `session_token` cookie is set (default 1 hour) which
        is required to submit a score when the session ends. Requires an active `user_token`
        cookie. The challenge must exist and be active.
      parameters:
      - description: ID of the challenge to begin a session for
        example: 1
        in: path
        name: challenge_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Session created successfully — session_token cookie is set
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid challenge ID, challenge does not exist, or missing
            user_token cookie
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — failed to generate session token or
            write to database
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a challenge session
      tags:
      - challenge_sessions
  /api/v1/challenges/session/end:
    patch:
      consumes:
      - application/json
      description: Marks an active challenge session as complete and saves the player's
        final score. This endpoint validates both the `user_token` and `session_token`
        cookies to prevent score manipulation. After saving the score, it automatically
        updates the leaderboard if the new score is the user's personal best. The
        `session_token` cookie is cleared on success. Requires both `user_token` and
        `session_token` cookies to be present.
      parameters:
      - description: The player's final score for the challenge attempt
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/database.Score'
      produces:
      - application/json
      responses:
        "200":
          description: Session ended and score saved successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing user_token or session_token cookie, or invalid score
            payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — failed to end session, save score,
            or update leaderboard
          schema:
            additionalProperties:
              type: string
            type: object
      summary: End a challenge session and submit score
      tags:
      - challenge_sessions
  /api/v1/leaderboard:
    get:
      description: Returns a ranked list of the top players for a specific challenge,
        ordered by best score descending. Each entry contains the player's username,
        their best score, and their Rocket League rank. Use this to display the challenge
        leaderboard and show where the current user stands relative to others.
      parameters:
      - description: Unique challenge ID to fetch the leaderboard for
        example: 1
        in: query
        name: challenge_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ranked leaderboard entries for the given challenge
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Invalid or non-numeric challenge_id provided
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get the leaderboard for a challenge
      tags:
      - leaderboards
  /api/v1/me/best-score:
    get:
      description: Returns the single highest score achieved by the authenticated
        user for the given challenge. This is the score used for the leaderboard.
        Returns null/empty if the user has not yet completed the challenge. Authentication
        is via the `user_token` HttpOnly cookie.
      parameters:
      - description: ID of the challenge to retrieve the best score for
        example: 1
        in: query
        name: challenge_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The user's personal best score entry for this challenge
          schema:
            $ref: '#/definitions/database.Score'
        "400":
          description: Missing user_token cookie or invalid challenge_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user's personal best score
      tags:
      - scores
  /api/v1/me/percentile:
    get:
      description: Calculates and returns the user's percentile position (0–100) on
        the global leaderboard for a specific challenge. A score of 95.0 means the
        user is in the top 5% of all players. A score of 50.0 means they are exactly
        at the median. Returns 0.0 if the user has not yet completed the challenge.
        Authentication is via the `user_token` HttpOnly cookie.
      parameters:
      - description: ID of the challenge to calculate percentile for
        example: 1
        in: query
        name: challenge_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'The user''s percentile ranking — e.g. {\"percentile\": 87.5}'
          schema:
            additionalProperties:
              format: float64
              type: number
            type: object
        "400":
          description: Missing user_token cookie or invalid challenge_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user's percentile ranking
      tags:
      - scores
  /api/v1/me/scores:
    get:
      description: Returns every recorded score for the authenticated user on a specific
        challenge, ordered by most recent first. Use this to display a score history
        graph or replay timeline. Authentication is via the `user_token` HttpOnly
        cookie. Requires a valid `challenge_id` query parameter.
      parameters:
      - description: ID of the challenge to retrieve scores for
        example: 1
        in: query
        name: challenge_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All recorded scores for this user on the given challenge
          schema:
            items:
              $ref: '#/definitions/database.Score'
            type: array
        "400":
          description: Missing user_token cookie or invalid challenge_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all scores for the current user
      tags:
      - scores
  /api/v1/rank:
    get:
      description: Returns a single Rocket League rank by its unique numeric ID. Returns
        404 if the rank does not exist. Rank IDs are stable — they are seeded in order
        from Bronze I (id=1) to Supersonic Legend (id=87).
      parameters:
      - description: Unique rank ID
        example: 51
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rank found
          schema:
            $ref: '#/definitions/database.Rank'
        "400":
          description: Invalid or non-numeric rank ID provided
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No rank found with the given ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a rank by ID
      tags:
      - ranks
  /api/v1/ranks:
    get:
      description: Returns the full list of all Rocket League ranks seeded in the
        database, ordered from Bronze I to Supersonic Legend. Use this to populate
        the rank selection dropdown on the user profile page. Each rank has a tier
        name (e.g. "Diamond"), a tier number (1–3), and a division (1–4).
      produces:
      - application/json
      responses:
        "200":
          description: Full list of Rocket League ranks
          schema:
            items:
              $ref: '#/definitions/database.Rank'
            type: array
        "500":
          description: Internal server error — database query failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all Rocket League ranks
      tags:
      - ranks
  /api/v1/users/guest:
    post:
      description: Instantly creates an anonymous guest account with no sign-up required.
        No request body needed. A random username (e.g. "Guest-a3f9c2") and a secure
        64-character hex identity token are generated automatically. The token is
        set as an HttpOnly `user_token` cookie valid for 7 days. All challenge scores
        are tied to this identity — calling this endpoint again creates a fresh guest
        account.
      produces:
      - application/json
      responses:
        "201":
          description: Guest account created — user_token cookie is set
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — failed to generate token or write to
            database
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a guest user account
      tags:
      - users
  /api/v1/users/me:
    patch:
      consumes:
      - application/json
      description: Allows a guest user to personalise their profile by setting a custom
        display name and their Rocket League rank. This converts a "basic guest" into
        an "improved guest" with a real identity on the leaderboard. Both fields are
        optional — send only the ones you want to update. Authentication is via the
        `user_token` HttpOnly cookie. The rank_id must correspond to a valid rank
        from GET /api/v1/ranks.
      parameters:
      - description: Profile fields to update
        in: body
        name: profile
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing user_token cookie or invalid JSON body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error — database update failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update guest user profile
      tags:
      - users
swagger: "2.0"